requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
authlib
jinja2
fastapi-mail
passlib[bcrypt]
python-jose
itsdangerous
email-validator
python-multipart
starlette
bcrypt
pydantic
requests
websockets
httpx
aiokafka
motor